src/translation_service/requirements.txt
# API and Web Framework
fastapi==0.109.1
uvicorn==0.34.0
python-multipart==0.0.6

# HTTP and Networking
requests==2.32.2
httpx==0.24.1

# HTML Processing
beautifulsoup4>=4.12.3
markdownify==0.14.1  # HTML to Markdown converter
html2text>=2020.1.16

# Data Validation
pydantic==2.10.6
pydantic-settings==2.8.0
python-dotenv==1.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Type Checking
mypy==1.7.1
pytest-cov==4.1.0
ruff==0.9.7